pyproject.toml
[project]
name = "yuzubot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "discord>=2.3.2",
    "jishaku>=2.6.3",
    "pillow>=12.1.1",
    "requests>=2.32.5",
]
